requirements.txt
numpy
sacrebleu==1.4.14
scikit-learn
torchvision
tqdm
